SolrSearch: Extract impl of findSolrByCollection
